Created loan registration screen

+++ app/src/main/java/me/isachenko/loansonline/data/model/requests/LoanRequest.kt
@@ -4,7 +4,7 @@ data class LoanRequest(
     val amount: Int,
     val firstName: String,
     val lastName: String,
-    val percent: Int,
+    val percent: Double,
     val period: Int,
     val phoneNumber: String
 )
